+++ backend/src/main/java/com/gabriel/minhacasa/domain/DTO/UpdateImmobileDTO.java
@@ -26,13 +26,13 @@ public record UpdateImmobileDTO(
         @NotNull
         int quantityRooms,
         @NotNull
-        float IPTU,
+        Double IPTU,
         @NotNull
-        float price,
+        Double price,
         @NotNull
         boolean suite,
         @NotNull
-        float totalArea,
+        Double totalArea,
         @NotNull
         int quantityBathrooms,
         @NotNull
